refactor: simplify raw workspace reads
